Drop the dtype knob, and the last of torch with it
Every caller in the repo passed dtype=np.float64, which was already the
default, so the parameter threaded five layers deep to do nothing. It existed
because operators used to follow torch.get_default_dtype, i.e. float32.

It was never fully wired anyway: _orthonormal_entries took a dtype and then
called orthonormalize_mappings without it, so the Gram matrix, its inverse
square root and the eigendecomposition all ran in float64 regardless, and the
dtype only cast the result afterwards.

Gone from get_reduction, get_coupling_operator, get_harmonic_operator,
get_reduction_of_weight, _orthonormal_entries, orthonormalize_mappings,
evaluate_tensors, tp_delta_epsilon and the two _contract_delta_epsilon
functions. dij() and eijk() are plain constants now: the test fixtures were
float32 only because the operators were, and with them in float64 remove_trace
no longer needs a delta that follows its caller's dtype.

Three tolerances were sized, and three comments argued, for a float32
evaluation that no longer happens. Both of the ones that mattered were measured
by tightening until they broke, rather than guessed:

  FINGERPRINT_RTOL  1e-5              -> 1e-10  (suite passes at 1e-12)
  OPERATOR_RTOL     1e-5, atol=1e-7   -> 1e-11, atol=0  (floor ~1e-13)

test_traceless_check_accepts_any_float_dtype is deleted. Its premise, that
np.allclose raises on a dtype mismatch, is false -- that was torch.allclose --
so it guarded a bug that can no longer occur. is_traceless keeps a plain test.

Also stale and removed: a TorchScript comment in coupling.py, and a TODO.md
paragraph and open item about accuracy criteria for float32 storage.

check_symmetry keeps its loose default tolerances. They apply to the caller's
own tensor rather than to anything natto evaluates, so tightening them would
change behaviour for existing callers; only the docstring is corrected.

+++ tests/test_orthonormal.py
@@ -22,7 +22,7 @@ RECONSTRUCTION_TOL = 1e-10
 
 def orthonormal(rank: int, symmetry: str = None, weight: int = None) -> dict:
     """The reduction in the self-dual basis, in double precision."""
-    output = get_reduction(rank, symmetry, basis="orthonormal", dtype=np.float64)
+    output = get_reduction(rank, symmetry, basis="orthonormal")
 
     return output if weight is None else output[weight]
